Begin validation and align backend with Front end
Changed band serializer to pick out ids of tasks and expenses
which aids in routing to a bands content on the front end.
Updated index method for tasks and expenses so it only
searches for those with a correlating band_id. Tested these changes
with curl requests. Additionally, added `:band_id` to task and expense
serializers