Update back end validation
Added validation to expenses, contacts, and tasks. Now for contacts
if you do not have a person's phone number, you can leave that
field blank and if you do have a number it will validate its length
and numericality. Did something similar for expense, if a user doesn't
have a cost at the moment they create it