Added current user to tasks, expenses, and contacts in the Controller
Re-defined the instance variable for each resource as well as changed
the create function so it correlates the creation with current user.
Made a direct connection in the user model to each resource with a
`has_many through` relationship. Also added `null: true` to each
resources migrate so they can't be created without a `band_id.`
Tested all of these with curl requests.